+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 2025-25-08
+### Workflow improvements and fixes
+- Pool emitters improvements
+- Audio refactor
+- Added icons for Scriptable Objects!
+
 ## [1.0.0] - 2025-07-02
 ### First Release
 - Include SFX, VFX controllers